Move SimplifyConvPad to a new pass and don't enable it by default
